# Conflicts:
#	doc/usage/configuration.rst
#	sphinx/directives/other.py
#	sphinx/environment/__init__.py
#	sphinx/environment/adapters/indexentries.py
#	sphinx/environment/adapters/toctree.py
#	sphinx/util/i18n.py
#	sphinx/util/nodes.py

+++ sphinx/util/fileutil.py
@@ -4,15 +4,21 @@ from __future__ import annotations
 
 import os
 import posixpath
-from typing import TYPE_CHECKING, Any, Callable
+from pathlib import Path
+from typing import TYPE_CHECKING
 
-from docutils.utils import relative_path
-
-from sphinx.util.osutil import copyfile, ensuredir
+from sphinx.locale import __
+from sphinx.util import logging
+from sphinx.util.osutil import _relative_path, copyfile, ensuredir
 
 if TYPE_CHECKING:
+    from collections.abc import Callable
+    from typing import Any
+
     from sphinx.util.template import BaseRenderer
     from sphinx.util.typing import PathMatcher
+
+logger = logging.getLogger(__name__)
 
 
 def _template_basename(filename: str | os.PathLike[str]) -> str | None:
@@ -28,9 +34,14 @@ def _template_basename(filename: str | os.PathLike[str]) -> str | None:
     return None
 
 
-def copy_asset_file(source: str | os.PathLike[str], destination: str | os.PathLike[str],
-                    context: dict[str, Any] | None = None,
-                    renderer: BaseRenderer | None = None) -> None:
+def copy_asset_file(
+    source: str | os.PathLike[str],
+    destination: str | os.PathLike[str],
+    context: dict[str, Any] | None = None,
+    renderer: BaseRenderer | None = None,
+    *,
+    force: bool = False,
+) -> None:
     """Copy an asset file to destination.
 
     On copying, it expands the template variables if context argument is given and
@@ -40,37 +51,70 @@ def copy_asset_file(source: str | os.PathLike[str], destination: str | os.PathLi
     :param destination: The path to destination file or directory
     :param context: The template variables.  If not given, template files are simply copied
     :param renderer: The template engine.  If not given, SphinxRenderer is used by default
+    :param bool force: Overwrite the destination file even if it exists.
     """
     if not os.path.exists(source):
         return
 
-    if os.path.isdir(destination):
+    destination = Path(destination)
+    if destination.is_dir():
         # Use source filename if destination points a directory
-        destination = os.path.join(destination, os.path.basename(source))
-    else:
-        destination = str(destination)
+        destination /= os.path.basename(source)
 
     if _template_basename(source) and context is not None:
         if renderer is None:
             from sphinx.util.template import SphinxRenderer
+
             renderer = SphinxRenderer()
 
         with open(source, encoding='utf-8') as fsrc:
-            destination = _template_basename(destination) or destination
-            with open(destination, 'w', encoding='utf-8') as fdst:
-                fdst.write(renderer.render_string(fsrc.read(), context))
+            template_content = fsrc.read()
+        rendered_template = renderer.render_string(template_content, context)
+
+        if not force and destination.exists() and template_content != rendered_template:
+            msg = __(
+                'Aborted attempted copy from rendered template %s to %s '
+                '(the destination path has existing data).'
+            )
+            logger.warning(
+                msg,
+                os.fsdecode(source),
+                os.fsdecode(destination),
+                type='misc',
+                subtype='copy_overwrite',
+            )
+            return
+
+        destination = _template_basename(destination) or destination
+        with open(destination, 'w', encoding='utf-8') as fdst:
+            msg = __('Writing evaluated template result to %s')
+            logger.info(
+                msg,
+                os.fsdecode(destination),
+                type='misc',
+                subtype='template_evaluation',
+            )
+            fdst.write(rendered_template)
     else:
-        copyfile(source, destination)
+        copyfile(source, destination, force=force)
 
 
-def copy_asset(source: str | os.PathLike[str], destination: str | os.PathLike[str],
-               excluded: PathMatcher = lambda path: False,
-               context: dict[str, Any] | None = None, renderer: BaseRenderer | None = None,
-               onerror: Callable[[str, Exception], None] | None = None) -> None:
+def copy_asset(
+    source: str | os.PathLike[str],
+    destination: str | os.PathLike[str],
+    excluded: PathMatcher = lambda path: False,
+    context: dict[str, Any] | None = None,
+    renderer: BaseRenderer | None = None,
+    onerror: Callable[[str, Exception], None] | None = None,
+    *,
+    force: bool = False,
+) -> None:
     """Copy asset files to destination recursively.
 
     On copying, it expands the template variables if context argument is given and
     the asset is a template file.
+
+    Use ``copy_asset_file`` instead to copy a single file.
 
     :param source: The path to source file or directory
     :param destination: The path to destination directory
@@ -78,21 +122,26 @@ def copy_asset(source: str | os.PathLike[str], destination: str | os.PathLike[st
     :param context: The template variables.  If not given, template files are simply copied
     :param renderer: The template engine.  If not given, SphinxRenderer is used by default
     :param onerror: The error handler.
+    :param bool force: Overwrite the destination file even if it exists.
     """
-    if not os.path.exists(source):
+    source = Path(source)
+    if not source.exists():
         return
 
     if renderer is None:
         from sphinx.util.template import SphinxRenderer
+
         renderer = SphinxRenderer()
 
     ensuredir(destination)
-    if os.path.isfile(source):
-        copy_asset_file(source, destination, context, renderer)
+    if source.is_file():
+        copy_asset_file(
+            source, destination, context=context, renderer=renderer, force=force
+        )
         return
 
     for root, dirs, files in os.walk(source, followlinks=True):
-        reldir = relative_path(source, root)
+        reldir = _relative_path(Path(root), source).as_posix()
         for dir in dirs.copy():
             if excluded(posixpath.join(reldir, dir)):
                 dirs.remove(dir)
@@ -102,9 +151,13 @@ def copy_asset(source: str | os.PathLike[str], destination: str | os.PathLike[st
         for filename in files:
             if not excluded(posixpath.join(reldir, filename)):
                 try:
-                    copy_asset_file(posixpath.join(root, filename),
-                                    posixpath.join(destination, reldir),
-                                    context, renderer)
+                    copy_asset_file(
+                        posixpath.join(root, filename),
+                        posixpath.join(destination, reldir),
+                        context=context,
+                        renderer=renderer,
+                        force=force,
+                    )
                 except Exception as exc:
                     if onerror:
                         onerror(posixpath.join(root, filename), exc)
